% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{add_diamond}
\alias{add_diamond}
\title{diamond_df}
\usage{
add_diamond(plot, data, fill = "grey20", colour = NA)
}
\arguments{
\item{plot}{a ggplot2}

\item{data}{a data.frame}

\item{fill}{ggplot2 fill value}
}
\description{
This function creates diamonds for moderators and summary statistics.
}
