% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-and-methods.R
\name{as.character.meta_ninja}
\alias{as.character.meta_ninja}
\title{meta_ninja as.character method}
\usage{
\method{as.character}{meta_ninja}(x, ...)
}
\arguments{
\item{x}{a meta_ninja}

\item{...}{extra arguments}
}
\description{
meta_ninja as.character method
}
