% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-and-methods.R
\name{as.data.frame.meta_ninja}
\alias{as.data.frame.meta_ninja}
\title{meta_ninja as.data.frame method}
\usage{
\method{as.data.frame}{meta_ninja}(x, row.names = NULL, optional = NULL, ...)
}
\arguments{
\item{x}{an R object}

\item{row.names}{NULL or a character vector}

\item{optional}{Rubbish}

\item{...}{extra arguments}
}
\description{
meta_ninja as.data.frame method
}
