% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunctions.R
\name{character_matrix}
\alias{character_matrix}
\title{character_matrix}
\usage{
character_matrix(x, levels = NULL, pattern = ",")
}
\arguments{
\item{x}{a character vector}

\item{levels}{a vector. Provides the level (and order) of colnames.}

\item{pattern}{pattern provided to str_split}
}
\description{
creates a predictor matrix when cells can contain multiple tags
}
