% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-and-methods.R
\name{coef.meta_ninja}
\alias{coef.meta_ninja}
\title{coef.meta_ninja}
\usage{
\method{coef}{meta_ninja}(object, ...)
}
\arguments{
\item{object}{the meta_ninja}

\item{...}{specific moderators to extract}
}
\description{
coef.meta_ninja
}
