% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conigave20}
\alias{conigave20}
\alias{conigrave20}
\title{conigrave20}
\format{
An object of class \code{escalc} (inherits from \code{data.frame}) with 104 rows and 48 columns.
}
\usage{
conigrave20
}
\description{
conigrave20
}
\author{
Conigrave et al
}
\keyword{datasets}
