% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{count_levels}
\alias{count_levels}
\title{count_levels}
\usage{
count_levels(y, v, cluster, factor, data)
}
\arguments{
\item{y}{a string}

\item{v}{a string}

\item{cluster}{a string}

\item{factor}{a string}

\item{data}{a data object}
}
\description{
counts levels in factors
}
