% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-functions.R
\name{describe_all_mods}
\alias{describe_all_mods}
\title{describe_all_mods}
\usage{
describe_all_mods(obj)
}
\arguments{
\item{obj}{the character names of an object}
}
\description{
describe_all_mods
}
