% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunctions.R
\name{factor_to_matrix}
\alias{factor_to_matrix}
\title{factor_to_matrix}
\usage{
factor_to_matrix(factor, col_names = NULL)
}
\arguments{
\item{factor}{a factor.}

\item{col_names}{a vector of colnames. If none given, will use factor levels.}
}
\description{
Converts a factor to a predictor matrix
}
