% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunctions.R
\name{final_matrix_singular}
\alias{final_matrix_singular}
\title{get_final_matrix}
\usage{
final_matrix_singular(model_call)
}
\arguments{
\item{model_call}{the call of a matrix}
}
\description{
Evaluates models to see if sufficient information is available to moderate.
}
