% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_nicely2.R
\name{format_nicely2}
\alias{format_nicely2}
\title{format_nicely2}
\usage{
format_nicely2(
  x,
  round = 2,
  transf = function(x) {     x },
  effect_name = NULL,
  transf_name = NULL,
  hide_insig = TRUE,
  escape_pc = FALSE,
  p_digits = 3,
  leading_zero = FALSE,
  ci_sep = ", ",
  include_i2 = FALSE
)
}
\arguments{
\item{x}{a meta_list object}

\item{round}{a scalar.}

\item{transf}{a function. If provided will transform effects and confidence intervals.}

\item{effect_name}{a string. If provided, will rename Estimate column with string provided.}

\item{transf_name}{a character string. If provided, will name the transformed column.}

\item{hide_insig}{a bool.}

\item{escape_pc}{a bool. If TRUE, \% symbols will be escaped in header, captions and notes.}

\item{p_digits}{a scalar. The number of digits to round p to.}

\item{leading_zero}{a bool. If TRUE, p-values will have leading zeros}

\item{ci_sep}{separator for confidence intervals}

\item{include_i2}{A bool, should i2 be included next to baseline?}
}
\description{
Formats a list of moderated meta3 objects into a data.frame which can be used as a table in a LaTex document
}
