% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{funnel_plot}
\alias{funnel_plot}
\title{ninjaFunnel}
\usage{
funnel_plot(model, xlab = "Estimate", ylab = "Standard Error",
  alpha = 0.5, font = "serif", CI = F)
}
\arguments{
\item{model}{an object belonging to the class 'meta_ninja'. These objects are created by the function 'meta3_moderation'.}

\item{xlab}{a character string. Label for the x-axis.}

\item{ylab}{a character string. Label for the y-axis.}

\item{alpha}{a scalar. Sets alpha transparency.}

\item{font}{a character string. Set's font family. Defaults to times new roman ('serif')}

\item{CI}{a bool. If true, error bars will be plotted. totally unnecesary, but kind of fun.}
}
\description{
This function is used for plotting funnel plots
}
