% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta3_exports.R
\name{get_matrix_long}
\alias{get_matrix_long}
\title{get_matrix_long}
\usage{
get_matrix_long(y, v, cluster, x, data, ...)
}
\arguments{
\item{y}{A vector of k studies of effect size.}

\item{v}{A vector of k studies of sampling variance.}

\item{cluster}{A vector of k characters or numbers indicating the clusters.}

\item{x}{A predictor or a k x m matrix of level-2 and level-3 predictors where m is the number of predictors.}

\item{data}{An optional data frame containing the variables in the model.}

\item{...}{A bin for additional arguments}
}
\description{
recreates the long matrix from meta3
}
