% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-functions.R
\name{get_model_value}
\alias{get_model_value}
\title{get_value}
\usage{
get_model_value(x, value, m = NULL)
}
\arguments{
\item{x}{a meta_ninja object}

\item{value}{the value to extract}

\item{m}{the moderator to extract}
}
\description{
Returns a value for models contained in meta_ninja
}
