% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-functions.R
\name{get_moderators}
\alias{get_moderators}
\title{get_moderators}
\usage{
get_moderators(meta_ninja, p = 0.05)
}
\arguments{
\item{meta_ninja}{a meta_ninja object}

\item{p}{p value with which to assess significance}
}
\description{
return moderators and significe status
}
