% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunctions.R
\name{get_new_matrix}
\alias{get_new_matrix}
\title{get_new_matrix}
\usage{
get_new_matrix(model_call, dat)
}
\arguments{
\item{model_call}{the call of a matrix}
}
\description{
Gets the final matrix to be used in models.
}
