% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunctions.R
\name{meta3_by_factor}
\alias{meta3_by_factor}
\title{meta3_by_factor}
\usage{
meta3_by_factor(y, v, cluster, factor, data, names = NULL,
  output.models = F)
}
\arguments{
\item{y}{a string. The name of the effect size variable}

\item{v}{a string. The name of the sampling variance variable}

\item{cluster}{a string. The name of the cluster variable}

\item{factor}{a string.}

\item{data}{an object The name of the data object}

\item{names}{character vector. If given, default names will be replaced.}

\item{output.models}{a bool. If true, models are provided in a list.}
}
\description{
Creates summary statistics for three level meta-analyses across a factor.
}
