% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunctions.R
\name{meta3_ninja}
\alias{meta3_ninja}
\title{meta3_ninja}
\usage{
meta3_ninja(call, moderators, binary_intercept = 0,
  continuous_intercept = NULL, remove_empty_slopes)
}
\arguments{
\item{call}{a call list}

\item{moderators}{a vector of moderators}

\item{binary_intercept}{a numeric. Constrain the intercept for matricies with binary elements}

\item{continuous_intercept}{a numeric. Constrain the intercept for matricies with continuous elements}

\item{remove_empty_slopes}{a bool. Inspects each matrix and removes empty slopes as neccesary}
}
\description{
This function streamlines the moderation process. Moderation strategy varies based on whether a numeric, or factoral variable is supplied.
}
