% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tssem_variables.R
\docType{data}
\name{model}
\alias{model}
\title{tssem1_table}
\format{
An object of class \code{tssem1REM} (inherits from \code{meta}) of length 16.
}
\usage{
model
}
\description{
Creates a table from the pooled correlation matrix from tssem1
}
\keyword{datasets}
