% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunctions.R
\name{moderation_instructions}
\alias{moderation_instructions}
\title{moderation_instructions}
\usage{
moderation_instructions(...)
}
\arguments{
\item{...}{moderation instructions}
}
\description{
Takes instructions, can then be provided to moderate
}
