% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderation_matrix_plot.R
\name{moderation_matrix}
\alias{moderation_matrix}
\title{moderation_matrix}
\usage{
moderation_matrix(
  ...,
  effect_size = "Effect size",
  moderators = NULL,
  null_value = 0,
  transf = NULL,
  leading_zero = TRUE
)
}
\arguments{
\item{...}{meta_ninja models all sharing a common data name}

\item{effect_size}{a string}

\item{moderators}{a list of moderators to include, order retained.}

\item{null_value}{a scalar indicating non-sigificance (where the dashed line will be drawn).}

\item{transf}{function to transform values}

\item{leading_zero}{when true, leading zeros are allowed on the x-axis}
}
\description{
Plots a moderation matrix
}
