% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-and-methods.R
\name{plot.meta_ninja}
\alias{plot.meta_ninja}
\title{meta_ninja plot method}
\usage{
\method{plot}{meta_ninja}(x, y, ...)
}
\arguments{
\item{x}{model to print}

\item{y}{not used.}

\item{...}{additional arguments passed to ninjaForest.}
}
\description{
meta_ninja plot method
}
