% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunctions.R
\name{prepare_args}
\alias{prepare_args}
\title{prepare_args}
\usage{
prepare_args(args)
}
\arguments{
\item{args}{some args}
}
\description{
Takes a list of arguments, puts contents in names when there are none
}
