% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-and-methods.R
\name{print.haiku}
\alias{print.haiku}
\title{haiku print}
\usage{
\method{print}{haiku}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{additional arguments. Not currently used.}
}
\description{
haiku print
}
