% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-and-methods.R
\name{print.meta_ninja}
\alias{print.meta_ninja}
\title{meta_ninja methods}
\usage{
\method{print}{meta_ninja}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{additional arguments. Not currently used.}
}
\description{
meta_ninja methods
}
