% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-and-methods.R
\name{print.ninja_summary}
\alias{print.ninja_summary}
\title{print.ninja_summary}
\usage{
\method{print}{ninja_summary}(object, ...)
}
\arguments{
\item{object}{ninja object}

\item{...}{other arguments}
}
\description{
print.ninja_summary
}
