% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_bias.r
\name{prop_stronger}
\alias{prop_stronger}
\title{prop_stronger}
\usage{
prop_stronger(
  m,
  q,
  transf = NULL,
  tail = "above",
  R = 500,
  method = "calibrated"
)
}
\description{
Wrapper for MetaUtility prop_stronger
}
