% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-and-methods.R
\name{removed_mod_note}
\alias{removed_mod_note}
\title{removed_mod_note}
\usage{
removed_mod_note(
  x,
  capitalise_first = T,
  case = c("default", "lower", "sentence")
)
}
\arguments{
\item{x}{a meta_ninja model}

\item{capitalise_first}{A bool. Set to true if you want to capitalise the first word}

\item{case}{one of "default", "lower" or "sentence"}
}
\description{
Generates text to describe the removal of moderators
}
