% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tssem_variables.R
\name{report_tssem2}
\alias{report_tssem2}
\title{report_tssem2}
\usage{
report_tssem2(
  x,
 
    pattern = "The model had {assessment} fit $\\\\chi^2$({df}) = {chi}; $p$ = {p}; RMSEA = {RMSEA}; SRMR = {SRMR}; TLI = {TLI}"
)
}
\arguments{
\item{x}{wls model}
}
\description{
Report fit of tssem2
}
