% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunctions.R
\name{string_meta3}
\alias{string_meta3}
\title{string_meta3}
\usage{
string_meta3(y, v, cluster, x = NULL, data, model.name = NA,
  intercept.constraints = NULL)
}
\arguments{
\item{y}{a string. The name of the effect size variable}

\item{v}{a string. The name of the sampling variance variable}

\item{cluster}{a string. The name of the cluster variable}

\item{x}{a string.}

\item{data}{an object. data object}

\item{model.name}{a string. The name of the model}

\item{intercept.constraints}{a numeric. Set the value of the intercept}
}
\description{
Allows three-level meta-analyses to be specified with strings
}
