% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-and-methods.R
\name{summary.meta_ninja}
\alias{summary.meta_ninja}
\title{meta_ninja summary method}
\usage{
\method{summary}{meta_ninja}(object, ...)
}
\arguments{
\item{object}{model to summarise}

\item{...}{additional arguments passed to format_nicely}
}
\description{
meta_ninja summary method
}
