% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3psm.R
\name{threePSM}
\alias{threePSM}
\title{threePSM}
\usage{
threePSM(model, ...)
}
\arguments{
\item{model}{the meta3 model}

\item{...}{arguments passed to weight::weightfunct}
}
\description{
Perform a weight-function model to test whether the pvalue of estimates affects population estimates
}
