% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-functions.R
\name{to_apa}
\alias{to_apa}
\title{to_apa}
\usage{
to_apa(x, caption, note, escape = F, ...)
}
\arguments{
\item{x}{a pretty_ninja}

\item{caption}{a character.}

\item{note}{a charcter.}

\item{escape}{a bool}

\item{...}{additional functions can be deliverd to papaja::apa_table}
}
\description{
to_apa
}
