% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tssem_variables.R
\name{tssem1_table}
\alias{tssem1_table}
\title{tssem1_table}
\usage{
tssem1_table(model)
}
\arguments{
\item{model}{the tssem stage 1 model}
}
\description{
Creates a table from the pooled correlation matrix from tssem1
}
