% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tssem_variables.R
\name{tssem2_table}
\alias{tssem2_table}
\title{tssem2_table}
\usage{
tssem2_table(
  wls,
  ...,
  transf = NULL,
  t.name = NULL,
  estimate = "Estimate",
  round = 2
)
}
\arguments{
\item{wls}{a wls model}

\item{...}{recode variable names}

\item{transf}{transform function to apply to results}

\item{t.name}{name for transformed results}

\item{estimate}{string, name for estimate variable}
}
\description{
Tabulate tssem2 regressions
}
