% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunctions.R
\name{extract_colnames}
\alias{extract_colnames}
\title{extract_colnames}
\usage{
extract_colnames(model, n, data, iteration, pred_matricies)
}
\arguments{
\item{model}{the model to extract from}

\item{n}{the number of names to extract}

\item{data}{data}

\item{iteration}{the index to find the predictor matrix}

\item{pred_matricies}{a list of predictor matricies}
}
\description{
Tries to determine column names of matricies
}
