% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-functions.R
\name{format_nicely}
\alias{format_nicely}
\title{format_nicely}
\usage{
format_nicely(x, round = 2, transform = NULL, t_name = NULL)
}
\arguments{
\item{x}{a meta_ninja}

\item{round}{a scalar.}

\item{transform}{a function. If provided will transform effects and confidence intervals.}

\item{t_name}{a character string. If provided, will name the transformed column.}
}
\description{
format_nicely
}
