% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-functions.R
\name{format_nicely}
\alias{format_nicely}
\title{format_nicely}
\usage{
format_nicely(
  x,
  round = 2,
  effect.name = NULL,
  transform = NULL,
  t.name = NULL,
  hide.insig = T,
  escape.pc = F,
  p_digits = 3,
  leading.zero = FALSE
)
}
\arguments{
\item{x}{a meta_ninja}

\item{round}{a scalar.}

\item{effect.name}{a string. If provided, will rename Estimate column with string provided.}

\item{transform}{a function. If provided will transform effects and confidence intervals.}

\item{t.name}{a character string. If provided, will name the transformed column.}

\item{hide.insig}{a bool.}

\item{escape.pc}{a bool. If TRUE, \% symbols will be escaped in header, captions and notes.}

\item{p_digits}{a scalar. The number of digits to round p to.}

\item{leading.zero}{a bool. If TRUE, p-values will have leading zeros}
}
\description{
format_nicely
}
