% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunctions.R
\name{meta3_moderation}
\alias{meta3_moderation}
\title{meta3_moderation}
\usage{
meta3_moderation(y, v, cluster, moderators, data, base = NULL,
  names = NULL)
}
\arguments{
\item{y}{a string. The name of yi}

\item{v}{a string. The name of vi}

\item{cluster}{a string. The name of the clustering variable e.g. 'study id'.}

\item{moderators}{a charcater vector. The names of the variable used for moderation}

\item{data}{an object.}

\item{base}{a meta3 model. Defaults to NULL. The baseline model is automatically calculated, but can also be supplied with this argument.}

\item{names}{a vector of strings. If supplied, the models can be renamed.}
}
\description{
This function streamlines the moderation process. Moderation strategy varies based on whether a numeric, or factoral variable is supplied.
}
