% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunctions.R
\name{meta3_moderation}
\alias{meta3_moderation}
\title{meta3_moderation}
\usage{
meta3_moderation(call, moderators)
}
\arguments{
\item{call}{a call list}

\item{moderators}{a vector of moderators}
}
\description{
This function streamlines the moderation process. Moderation strategy varies based on whether a numeric, or factoral variable is supplied.
}
