% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunctions.R
\name{moderate}
\alias{moderate}
\title{moderate}
\usage{
moderate(model, ...)
}
\arguments{
\item{model}{A meta3 model. The original data file must be available in the environment, with the same name.}

\item{...}{moderators, entered as objects}
}
\description{
This is a wrapper to perform meta3 moderations with. The original data file must be in the environment.
}
