% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunctions.R
\name{moderate}
\alias{moderate}
\title{moderate}
\usage{
moderate(
  model,
  ...,
  moderators = NULL,
  binary_intercept = 0,
  continuous_intercept = NULL,
  remove_empty_slopes = T
)
}
\arguments{
\item{model}{A meta3 model. The original data file must be available in the environment, with the same name.}

\item{...}{moderators, entered as objects}

\item{moderators}{a character vector. A vector of moderator names may be supplied.}

\item{binary_intercept}{a numeric. Constrain the intercept for matricies with binary elements}

\item{continuous_intercept}{a numeric. Constrain the intercept for matricies with continuous elements}

\item{remove_empty_slopes}{a bool. If true, removes empty columns from matricies.}
}
\value{
a meta_ninja \(until I rename it\)
}
\description{
This is a wrapper to perform meta3 moderations with. The original data file must be in the environment.
}
\details{
moderate simplifies moderation analyses by taking the call from a meta3, and then using it to generate
subsequent moderation models. A few rules are used to do this.
1. If a continuous variable is used a predictor, then an intercept is recorded
2. If binary variables are included, then intercepts are forced to be zero, these binary variables become the intercepts.
}
