% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{ninjaForest}
\alias{ninjaForest}
\title{extractSlopes}
\usage{
ninjaForest(model, summary = NULL, save = F, height = 16,
  width = 16, f_name = test.png, title = "")
}
\arguments{
\item{model}{a model.}
}
\description{
Grabs slope coefficient data
}
