% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{ninjaForest}
\alias{ninjaForest}
\title{ninjaForest}
\usage{
ninjaForest(model, xlab = "Effect size", transform = NULL,
  baseline_name = "All", factor.levels = NULL, vline = NULL,
  cluster = NULL, author = NULL, year = NULL, moderator.shape = 23,
  moderator.size = 3.2, summary.shape = 23, summary.size = 4,
  font = "serif")
}
\arguments{
\item{model}{an object belonging to the class 'meta_ninja'. These objects are created by the function 'meta3_moderation'.}

\item{xlab}{a string.}

\item{transform}{a function. If supplied effect sizes are transformed by this function}

\item{baseline_name}{a string. The label for the baseline model}

\item{factor.levels}{a charcater vector. If supplied, only the factor.levels specified will be plotted.}

\item{vline}{a scalar. Dictates the x-intercept (the dashed line).}

\item{cluster}{study_id column}

\item{author}{the name of the author column}

\item{year}{the name of the year column}

\item{moderator.shape}{a scalar ggplot2 geom_point shape value}

\item{moderator.size}{a scalar. ggplot2 geom_point size value}

\item{font}{A string. The name of a font family. Defaults to serif.}

\item{pooled.shape}{a scalar ggplot2 geom_point shape value}

\item{pooled.size}{a scalar ggplot2 geom_point size value}
}
\description{
This function is used for plotting
}
