% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunctions.R
\name{split_to_matrix}
\alias{split_to_matrix}
\title{split_to_matrix}
\usage{
split_to_matrix(x, pattern = ",")
}
\arguments{
\item{x}{a character vector}

\item{pattern}{pattern provided to str_split}
}
\description{
creates a predictor matrix when cells can contain multiple tags
}
