% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-functions.R
\name{to_apa}
\alias{to_apa}
\title{to_apa}
\usage{
to_apa(x, caption, note, escape = F, escape.pc = FALSE, docx = T, ...)
}
\arguments{
\item{x}{a pretty_ninja}

\item{caption}{a character.}

\item{note}{a charcter.}

\item{escape}{a bool. Sent to papaja::apa_table, defaults to false.}

\item{escape.pc}{a bool. If True, \% symbols will be escaped in header, captions and notes.}

\item{docx}{a bool. If True, formats table superscripts for docx, otherwise for pdf.}

\item{...}{additional functions can be deliverd to papaja::apa_table}
}
\description{
to_apa
}
